refactor: component 구조 common 추가
