+++ src/layouts/layout.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 
-import Footer from '@/components/Footer/Footer';
-import Header from '@/components/Header/Header';
+import Footer from '@/components/common/Footer/Footer';
+import Header from '@/components/common/Header/Header';
 
 interface Props {
   children: ReactNode;
